% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simReadCounts}
\alias{simReadCounts}
\title{Simulate a very simple raw scRNA-seq expression matrix for testing}
\usage{
simReadCounts(nGenes, nCells, sparsity = 0.9, doublet = 0, lowQual = 0)
}
\arguments{
\item{nGenes}{The number of genes (rows) in the matrix}

\item{nCells}{The number of cells (columns) in the matrix}

\item{sparsity}{Default = 0.9. The sparsity of the matrix (percentage of 0's)}

\item{doublet}{Default = 0. The percentage of doublets}

\item{lowQual}{Default = 0. The percentage of low quality cells}
}
\value{
genes by cell matrix with ransom read counts
}
\description{
Simulate a very simple raw scRNA-seq expression matrix for testing
}
