% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simReadCounts}
\alias{simReadCounts}
\title{Simulate a very simple metagenomic taxa count matrix for testing purposes}
\usage{
simReadCounts(nTaxa, nSamples, sparsity, monoDom = 0, lowQual = 0)
}
\arguments{
\item{nTaxa}{The number of taxa (rows) in the matrix}

\item{nSamples}{The number of samples (columns) in the matrix}

\item{sparsity}{Default = 0.9. The sparsity of the matrix (percentage of 0's)}

\item{monoDom}{Default = 0. The percentage of samples with monoDom}

\item{lowQual}{Default = 0. The percentage of low quality samples}
}
\value{
taxa by sample matrix with random read counts
}
\description{
Simulate a very simple metagenomic taxa count matrix for testing purposes
}
